Adjust ClusterConfiguration
